fixed bugs in put and patch methods

+++ src/main/java/dev/learnSpring/basicOperations/TodoServices.java
@@ -19,7 +19,7 @@ public class TodoServices {
     }
 
     public Todo getTodoById(Long id){
-        return todoRepository.getReferenceById(id);
+        return todoRepository.findById(id).orElse(null);
     }
 
     public List<Todo> getAllTodos(){
@@ -27,9 +27,43 @@ public class TodoServices {
     }
 
     public Todo deleteTodoById(Long id){
-        Todo data = todoRepository.getReferenceById(id);
+        Todo data = todoRepository.findById(id).orElse(null);
+        if(data == null){
+            return null;
+        }
         todoRepository.delete(data);
         return data;
+    }
+
+
+
+    public Todo updateTodo(Long id ,Todo data){
+
+        Todo oData = todoRepository.findById(id).orElse(null);
+        if(oData == null){
+            return null;
+        }
+        oData.setTitle(data.getTitle());
+        oData.setDescription(data.getDescription());
+        oData.setStatus(data.getStatus());
+        return todoRepository.save(oData);
+    }
+
+    public Todo patchTodo(Long id , Todo data){
+        Todo oData = todoRepository.findById(id).orElse(null);
+        if(oData == null){
+            return null;
+        }
+
+        if(data.getTitle() != null){
+            oData.setTitle(data.getTitle());
+        }if(data.getDescription() != null){
+            oData.setDescription(data.getDescription());
+        }if(data.getStatus() != null){
+            oData.setStatus(data.getStatus());
+        }
+
+        return todoRepository.save(oData);
     }
